ppe/requirements.txt
torch==1.9.0
opencv_python==4.5.3.56
yolov5==5.0.10
Pillow==8.4.0
